fix(analytics): E2.2 ranking materializa COUNT(*) OVER() como bigint/long
O endpoint GET /api/insights/popularity/top dava 500 em TODA chamada: no
PostgreSQL COUNT(*) OVER() devolve bigint (Int64), mas RankingRow.TotalCount
era int (Int32), entao o Dapper nao achava construtor compativel e a
materializacao estourava em runtime. A query de summary funcionava so porque
o result dela ja usa long.

- RankingRow.TotalCount passa a long; o total e reduzido a int para o
  PagedResult no handler.
- Comentario fixando a regra para as proximas queries paginadas (E2.5+).

Encontrado por smoke test com Postgres real (Docker): os testes de contrato
do E2.2 passavam por afirmarem so o TEXTO do SQL, sem materializar a linha —
esse erro de tipo so aparece contra um banco vivo.

+++ src/Modules/Analytics/SpotifyDataAnalysis.Modules.Analytics.Application/Insights/GetPopularityRankingQuery.cs
@@ -109,7 +109,7 @@ internal sealed class GetPopularityRankingQueryHandler
 
         IReadOnlyList<RankingRow> rows = (await connection.QueryAsync<RankingRow>(command)).AsList();
 
-        int totalCount = rows.Count > 0 ? rows[0].TotalCount : 0;
+        int totalCount = rows.Count > 0 ? (int)rows[0].TotalCount : 0;
 
         IReadOnlyList<PopularityRankingItem> items = rows
             .Select(row => new PopularityRankingItem(
@@ -121,11 +121,16 @@ internal sealed class GetPopularityRankingQueryHandler
 
     // Linha crua do Dapper: o item de ranking + o TotalCount do COUNT(*) OVER() (repetido em cada linha da
     // página). Materializado no handler para o PagedResult carregar o total num só round-trip.
+    //
+    // TotalCount é `long` (não `int`) DE PROPÓSITO: no PostgreSQL, COUNT(*) OVER() devolve `bigint` (Int64), e
+    // o Dapper casa o construtor do record pelo tipo CLR da coluna retornada — um `int` aqui faz a
+    // materialização falhar em runtime ("no matching constructor", só visível contra um banco real). O total
+    // é reduzido a `int` para o PagedResult no handler. Toda query paginada nova (E2.5+) deve seguir isto.
     private sealed record RankingRow(
         string TrackId,
         string Name,
         string? PrimaryArtist,
         int Popularity,
         string? Genre,
-        int TotalCount);
+        long TotalCount);
 }
